Add all zeros and test
